Refactor index module
