+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod book;
 pub mod config;
-pub mod index;
+pub mod es;
 pub mod web;
